third_party/blink/web_tests/external/wpt/css/css-page/parsing/size-001-expected.txt

This is a testharness.js-based test.
FAIL size-001 assert_equals: for rule 7 expected "letter;" but got "letter portrait;"
Harness: the test ran to completion.
